Move redux utils into the newly created utils folder
