+++ app/ui/shared/util/redux/middleware/history.js
@@ -10,11 +10,10 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 */
 
-import rootReducer from '../reducers';
-import * as model from '../model/index';
-import configureStore from '../../shared/util/redux/configure-store';
+export default historyStore => _store => next => action => {
+  if (typeof action !== 'function') {
+    historyStore.push(action);
+  }
 
-export function createContentStore() {
-  const initialState = new model.State(rootReducer(undefined, { type: null }));
-  return configureStore(rootReducer, initialState);
-}
+  return next(action);
+};
